requirements.txt
flask
transformers
torch
Pillow
nltk>=3.8.1
pyOpenSSL